Replace OpenStudio 301EnergyRatingIndexRuleset measure with a simpler ruby script.
