+++ .github/workflows/config.yml
@@ -34,7 +34,7 @@ jobs:
         shell: bash
         run: |
           export PYTHONIOENCODING=utf8
-          export SCTPATH=rulesets/301EnergyRatingIndexRuleset/resources/301validator.xml
+          export SCTPATH=rulesets/resources/301validator.xml
           for xmlfile in workflow/sample_files/*.xml; do XMLPATH=$xmlfile python3 -c "import os; import sys; from lxml import isoschematron, etree; from pathlib import Path; sct = Path(os.environ.get('SCTPATH')).read_text(); sct_doc = etree.fromstring(bytes(sct, encoding='utf-8')); schematron = isoschematron.Schematron(sct_doc, store_report = True); xml_path = os.environ.get('XMLPATH'); xml = Path(xml_path).read_text(); xml_doc = etree.fromstring(bytes(xml, encoding='utf-8')); valid = schematron.validate(xml_doc); (not valid) and print(schematron.validation_report); (not valid) and print(f'{xml_path} fails to validate'); valid and print(f'{xml_path} validates'); (not valid) and sys.exit(1)"; done
 
       - name: Run measure unit tests
