--- rulesets/301EnergyRatingIndexRuleset/tests/test_es_pv_batteries.rb
@@ -1,87 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../../../hpxml-measures/HPXMLtoOpenStudio/resources/minitest_helper'
-require 'openstudio'
-require 'openstudio/ruleset/ShowRunnerOutput'
-require_relative '../measure.rb'
-require 'fileutils'
-require_relative 'util'
-
-class EnergyStarPVTest < MiniTest::Test
-  def setup
-    @root_path = File.absolute_path(File.join(File.dirname(__FILE__), '..', '..', '..'))
-    @tmp_hpxml_path = File.join(@root_path, 'workflow', 'sample_files', 'tmp.xml')
-  end
-
-  def teardown
-    File.delete(@tmp_hpxml_path) if File.exist? @tmp_hpxml_path
-  end
-
-  def test_pv
-    ESConstants.AllVersions.each do |es_version|
-      _convert_to_es('base-pv.xml', es_version)
-      hpxml = _test_measure()
-      _check_pv(hpxml)
-    end
-  end
-
-  def test_pv_batteries
-    skip # Temporarily disabled until RESNET allows this.
-    ESConstants.AllVersions.each do |es_version|
-      _convert_to_es('base-pv-battery.xml', es_version)
-      hpxml = _test_measure()
-      _check_battery(hpxml)
-    end
-  end
-
-  def _test_measure()
-    args_hash = {}
-    args_hash['hpxml_input_path'] = @tmp_hpxml_path
-    args_hash['init_calc_type'] = ESConstants.CalcTypeEnergyStarReference
-
-    # create an instance of the measure
-    measure = EnergyRatingIndex301Measure.new
-
-    # create an instance of a runner
-    runner = OpenStudio::Measure::OSRunner.new(OpenStudio::WorkflowJSON.new)
-
-    model = OpenStudio::Model::Model.new
-
-    # get arguments
-    arguments = measure.arguments(model)
-    argument_map = OpenStudio::Measure.convertOSArgumentVectorToMap(arguments)
-
-    # populate argument with specified hash value if specified
-    arguments.each do |arg|
-      temp_arg_var = arg.clone
-      if args_hash.has_key?(arg.name)
-        assert(temp_arg_var.setValue(args_hash[arg.name]))
-      end
-      argument_map[arg.name] = temp_arg_var
-    end
-
-    # run the measure
-    measure.run(model, runner, argument_map)
-    result = runner.result
-
-    # show the output
-    show_output(result) unless result.value.valueName == 'Success'
-
-    # assert that it ran correctly
-    assert_equal('Success', result.value.valueName)
-
-    return measure.new_hpxml
-  end
-
-  def _check_pv(hpxml)
-    assert_equal(0, hpxml.pv_systems.size)
-  end
-
-  def _check_battery(hpxml)
-    assert_equal(0, hpxml.batteries.size)
-  end
-
-  def _convert_to_es(hpxml_name, program_version, state_code = nil)
-    return convert_to_es(hpxml_name, program_version, @root_path, @tmp_hpxml_path, state_code)
-  end
-end
