+++ rulesets/main.rb
@@ -1,0 +1,186 @@
+# frozen_string_literal: true
+
+require 'pathname'
+require 'csv'
+require 'oga'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/airflow'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/battery'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/constants'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/constructions'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/geometry'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/hotwater_appliances'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/hpxml'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/hpxml_defaults'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/hvac'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/hvac_sizing'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/lighting'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/materials'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/misc_loads'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/psychrometrics'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/schedules'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/unit_conversions'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/util'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/validator'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/waterheater'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/weather'
+require_relative '../hpxml-measures/HPXMLtoOpenStudio/resources/xmlhelper'
+require_relative 'resources/301ruleset'
+require_relative 'resources/ESruleset'
+require_relative 'resources/constants'
+
+def run_rulesets(runner, hpxml_input_path, designs)
+  unless (Pathname.new hpxml_input_path).absolute?
+    hpxml_input_path = File.expand_path(File.join(File.dirname(__FILE__), hpxml_input_path))
+  end
+  unless File.exist?(hpxml_input_path) && hpxml_input_path.downcase.end_with?('.xml')
+    runner.registerError("'#{hpxml_input_path}' does not exist or is not an .xml file.")
+    return false
+  end
+
+  begin
+    stron_paths = []
+    stron_paths << File.join(File.dirname(__FILE__), '..', 'hpxml-measures', 'HPXMLtoOpenStudio', 'resources', 'hpxml_schematron', 'HPXMLvalidator.xml')
+    stron_paths << File.join(File.dirname(__FILE__), 'resources', '301validator.xml')
+    orig_hpxml = HPXML.new(hpxml_path: hpxml_input_path, schematron_validators: stron_paths)
+    orig_hpxml.errors.each do |error|
+      runner.registerError(error)
+    end
+    orig_hpxml.warnings.each do |warning|
+      runner.registerWarning(warning)
+    end
+    return false unless orig_hpxml.errors.empty?
+
+    # Weather file
+    epw_path = orig_hpxml.climate_and_risk_zones.weather_station_epw_filepath
+    if not File.exist? epw_path
+      test_epw_path = File.join(File.dirname(hpxml_input_path), epw_path)
+      epw_path = test_epw_path if File.exist? test_epw_path
+    end
+    if not File.exist? epw_path
+      test_epw_path = File.join(File.dirname(__FILE__), '..', 'weather', epw_path)
+      epw_path = test_epw_path if File.exist? test_epw_path
+    end
+    if not File.exist?(epw_path)
+      runner.registerError("'#{epw_path}' could not be found.")
+      return false
+    end
+
+    cache_path = epw_path.gsub('.epw', '-cache.csv')
+    if not File.exist?(cache_path)
+      runner.registerError("'#{cache_path}' could not be found. Perhaps you need to run: openstudio energy_rating_index.rb --cache-weather")
+      return false
+    end
+
+    # Obtain weather object
+    weather = WeatherProcess.new(nil, nil, cache_path)
+
+    eri_version = orig_hpxml.header.eri_calculation_version
+    eri_version = Constants.ERIVersions[-1] if eri_version == 'latest'
+    if not eri_version.nil?
+      # Obtain egrid subregion & cambium gea region
+      egrid_subregion = get_epa_egrid_subregion(runner, orig_hpxml)
+      if Constants.ERIVersions.index(eri_version) >= Constants.ERIVersions.index('2019ABCD')
+        cambium_gea = get_cambium_gea_region(runner, orig_hpxml)
+      end
+    end
+
+    create_time = Time.now.strftime('%Y-%m-%dT%H:%M:%S%:z')
+
+    last_hpxml = nil
+    new_hpxml_strings = {}
+    designs.each do |design|
+      # Ensure we don't modify the original HPXML
+      new_hpxml = Marshal.load(Marshal.dump(orig_hpxml))
+
+      # Apply initial ruleset on HPXML object
+      if [ESConstants.CalcTypeEnergyStarReference,
+          ESConstants.CalcTypeEnergyStarRated].include? design.init_calc_type
+        new_hpxml = EnergyStarRuleset.apply_ruleset(new_hpxml, design.init_calc_type)
+      end
+
+      # Write initial HPXML file
+      if not design.init_hpxml_output_path.nil?
+        if not File.exist? design.init_hpxml_output_path
+          XMLHelper.write_file(new_hpxml.to_oga, design.init_hpxml_output_path)
+          runner.registerInfo("Wrote file: #{design.init_hpxml_output_path}")
+        end
+      end
+
+      # Apply 301 ruleset on HPXML object
+      if not design.calc_type.nil?
+        new_hpxml = EnergyRatingIndex301Ruleset.apply_ruleset(runner, new_hpxml, design.calc_type, weather,
+                                                              design.iecc_version, egrid_subregion, cambium_gea, create_time)
+      end
+      last_hpxml = new_hpxml
+
+      # Write final HPXML file
+      if not design.hpxml_output_path.nil?
+        new_hpxml_strings[design.hpxml_output_path] = XMLHelper.write_file(new_hpxml.to_oga, design.hpxml_output_path)
+        runner.registerInfo("Wrote file: #{design.hpxml_output_path}")
+      end
+    end
+  rescue Exception => e
+    runner.registerError("#{e.message}\n#{e.backtrace.join("\n")}")
+    return false
+  end
+
+  duplicates = {}
+  new_hpxml_strings.each_with_index do |(hpxml_output_path, new_hpxml), i|
+    next if i == 0
+
+    new_hpxml_strings.each_with_index do |(hpxml_output_path2, new_hpxml2), j|
+      next if j >= i
+
+      if new_hpxml == new_hpxml2
+        duplicates[hpxml_output_path] = hpxml_output_path2
+        break
+      end
+    end
+  end
+
+  return true, duplicates, last_hpxml
+end
+
+def get_epa_egrid_subregion(runner, hpxml)
+  egrid_zip_filepath = File.join(File.dirname(__FILE__), 'data', 'egrid', 'ZIP_mappings.csv')
+  egrid_subregion = lookup_region_from_zip(hpxml.header.zip_code, egrid_zip_filepath, 0, 1)
+  if egrid_subregion.nil?
+    runner.registerWarning("Could not look up eGRID subregion for zip code: '#{hpxml.header.zip_code}'. Emissions will not be calculated.")
+  end
+  return egrid_subregion
+end
+
+def get_cambium_gea_region(runner, hpxml)
+  cambium_zip_filepath = File.join(File.dirname(__FILE__), 'data', 'cambium', 'ZIP_mappings.csv')
+  cambium_gea = lookup_region_from_zip(hpxml.header.zip_code, cambium_zip_filepath, 0, 1)
+  if cambium_gea.nil?
+    runner.registerWarning("Could not look up Cambium GEA for zip code: '#{hpxml.header.zip_code}'. CO2e emissions will not be calculated.")
+  end
+  return cambium_gea
+end
+
+def lookup_region_from_zip(zip_code, zip_filepath, zip_column_index, output_column_index)
+  return if zip_code.nil?
+
+  if zip_code.include? '-'
+    zip_code = zip_code.split('-')[0]
+  end
+  zip_code = zip_code.rjust(5, '0')
+
+  return if zip_code.size != 5
+
+  begin
+    Integer(zip_code)
+  rescue
+    return
+  end
+
+  CSV.foreach(zip_filepath) do |row|
+    fail "Zip code in #{zip_filepath} needs to be 5 digits." if zip_code.size != 5
+    next unless row[zip_column_index] == zip_code
+
+    return row[output_column_index]
+  end
+
+  return
+end
