+++ rulesets/resources/ESruleset.rb
