+++ rulesets/resources/constants.rb
