+++ tasks.rb
@@ -2594,7 +2594,7 @@ if ARGV[0].to_sym == :update_measures
   require_relative 'hpxml-measures/HPXMLtoOpenStudio/resources/hpxml'
   require_relative 'hpxml-measures/HPXMLtoOpenStudio/resources/lighting'
   require_relative 'hpxml-measures/HPXMLtoOpenStudio/resources/xmlhelper'
-  require_relative 'rulesets/301EnergyRatingIndexRuleset/resources/constants'
+  require_relative 'rulesets/resources/constants'
 
   # Prevent NREL error regarding U: drive when not VPNed in
   ENV['HOME'] = 'C:' if !ENV['HOME'].nil? && ENV['HOME'].start_with?('U:')
@@ -2672,8 +2672,7 @@ if ARGV[0].to_sym == :create_release_zips
            'hpxml-measures/HPXMLtoOpenStudio/resources/**/*.*',
            'hpxml-measures/ReportSimulationOutput/measure.*',
            'hpxml-measures/ReportSimulationOutput/resources/**/*.*',
-           'rulesets/301EnergyRatingIndexRuleset/measure.*',
-           'rulesets/301EnergyRatingIndexRuleset/resources/**/*.*',
+           'rulesets/**/*.*',
            'weather/*.*',
            'workflow/*.*',
            'workflow/real_homes/*.*',
