+++ workflow/tests/real_homes_test.rb
@@ -7,7 +7,7 @@ require 'fileutils'
 require 'csv'
 require 'oga'
 require_relative 'util.rb'
-require_relative '../../rulesets/301EnergyRatingIndexRuleset/resources/constants'
+require_relative '../../rulesets/resources/constants'
 require_relative '../../hpxml-measures/HPXMLtoOpenStudio/resources/constants'
 require_relative '../../hpxml-measures/HPXMLtoOpenStudio/resources/hpxml'
 require_relative '../../hpxml-measures/HPXMLtoOpenStudio/resources/xmlhelper'
